l4: add udp port handler
For VxLan tunnels, we need to register a specific destination udp port.

Add a new node and api, and push all non-handled packets back
to the management (loopback) interface, as done today.
